feat: check if user owns asset before deleting asset

+++ packages/api/src/plugins/assets/handlers/deleteAssetHandler.ts
@@ -1,21 +1,26 @@
 import Boom from '@hapi/boom';
 import Hapi from '@hapi/hapi';
+import { checkIfUserOwnsAsset } from '../../../utils/checkIfUserOwnsAsset';
 import { handlePrismaError } from '../../../utils/handlePrismaError';
 import { AssetInput } from '../assets';
 
-
 export async function deleteAssetHandler(request: Hapi.Request, h: Hapi.ResponseToolkit) {
   const { prisma } = request.server.app;
-  const payload = request.payload as AssetInput;
-  const username = request.auth.credentials.userName as string;
+  const { id } = request.payload as AssetInput;
+  const userId = request.auth.credentials.userId as number;
 
   try {
-    await prisma.asset.delete({
-      where: {
-        id: payload.id,
-      },
-    });
-    return h.response().code(204);
+    const isAllowed = await checkIfUserOwnsAsset(request, id, userId);
+    if (isAllowed) {
+      await prisma.asset.delete({
+        where: {
+          id,
+        },
+      });
+      return h.response().code(204);
+    } else {
+      throw new Error('You are not allowed');
+    }
   } catch (err) {
     console.log(err);
     return handlePrismaError(err);
